add: add Stripe Connect feature and add business logic to store Transaction data

+++ src/main/java/com/wequan/bu/repository/model/Tutor.java
@@ -24,5 +24,5 @@ public class Tutor {
     private String currentCity;
     private String currentState;
     private String teachMethod;
-    private Float hourlyRate;
+    private Integer hourlyRate;
 }
