+++ src/main/java/com/wequan/bu/repository/model/TutorApplication.java
@@ -23,7 +23,7 @@ public class TutorApplication {
     private String currentCity;
     private String currentState;
     private String teachMethod;
-    private Float hourlyRate;
+    private Integer hourlyRate;
     private Short latePolicyId;
     private Short cancellationPolicyId;
     /**
